-add forward pruning based on num of children -fix heuristic size issue

+++ gameAI.h
@@ -12,4 +12,4 @@ bool checkWin(vector<vector<int> > currBoard, int wCaps, int bCaps, int color, i
 pair<vector<pii>, vector<pii> > getChildren(vector<vector<int> > currBoard);
 priority_queue <pair<int, pii> > moveOrderingMax(vector<pii> children, vector<vector<int> > board, int agentTile, int depth, int alpha, int beta, bool isMaximizing, int wCaps, int bCaps, ZobristHash zobrist);
 priority_queue <pair<int, pii>, vector<pair<int, pii> >, Compare > moveOrderingMin(vector<pii> children, vector<vector<int> > board, int agentTile, int depth, int alpha, int beta, bool isMaximizing, int wCaps, int bCaps, ZobristHash zobrist);
-int alphaBeta(vector<vector<int> > currBoard, int wCaps, int bCaps, int depth, int alpha, int beta, bool isMaximizing, int agentTile, ZobristHash hasher);
+long long int alphaBeta(vector<vector<int> > currBoard, int wCaps, int bCaps, int depth, long long int alpha, long long int beta, bool isMaximizing, int agentTile, ZobristHash hasher);
